separated LPL into the helper file, fixed no updating when in post-game lobby.
